*LIttle bit of cleanup, fix a warning.
--HG--
branch : trunk

+++ src/game/Map.cpp
@@ -217,12 +217,12 @@ Map::Map(uint32 id, time_t expiry, uint32 InstanceId, uint8 SpawnMode, Map* _par
 {
     m_notifyTimer.SetInterval(IN_MILISECONDS/2);
 
-    for (unsigned int idx=0; idx < MAX_NUMBER_OF_GRIDS; ++idx)
+    for (uint8 idx = 0; idx < MAX_NUMBER_OF_GRIDS; ++idx)
     {
-        for (unsigned int j=0; j < MAX_NUMBER_OF_GRIDS; ++j)
+        for (uint8 j = 0; j < MAX_NUMBER_OF_GRIDS; ++j)
         {
             //z code
-            GridMaps[idx][j] =NULL;
+            GridMaps[idx][j] = NULL;
             setNGrid(NULL, idx, j);
         }
     }
